+++ Gw2Sharp/WebApi/V2/Models/Skills/TransformationForm.cs
@@ -1,0 +1,18 @@
+namespace Gw2Sharp.WebApi.V2.Models
+{
+    /// <summary>
+    /// A form.
+    /// </summary>
+    public enum TransformationForm
+    {
+        /// <summary>
+        /// Unknown form.
+        /// </summary>
+        Unknown,
+
+        /// <summary>
+        /// Celestial avatar form.
+        /// </summary>
+        CelestialAvatar
+    }
+}