+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Gw2Sharp History
 
+## 0.9.5
+### Endpoints
+- Add the following missing properties in `Gw2Sharp.WebApi.V2.Models.Skill`:
+  - `Specialization`
+  - `DualAttunement`
+  - `SubSkills`
+- Change the following properties from `int` to `double`:
+  - `Gw2Sharp.WebApi.V2.Models.SkillFactComboFinisher.Percent`
+  - `Gw2Sharp.WebApi.V2.Models.SkillFactPercent.Percent`
+  - `Gw2Sharp.WebApi.V2.Models.SkillFactRecharge.Value`
+- Add missing property `Value` in `Gw2Sharp.WebApi.V2.Models.SkillFactPercent` (although this seems to be a bug in the API)
+- Deprecate `Gw2Sharp.WebApi.V2.Models.SkillWeaponType.Harpoon` in favor of `Spear` (still no consistency between harpoons, spears, harpoon guns and spear guns)
+- Add missing enum `None` in `Gw2Sharp.WebApi.V2.Models.AttributeType` (which is apparently used for skills that apply barrier)
+- Add missing enums `Toolbelt` and `Transform1` in `Gw2Sharp.WebApi.V2.Models.SkillSlot`
+- Add missing enum `Transform` in `Gw2Sharp.WebApi.V2.Models.SkillType`
+
 ## 0.9.4
 ### Fixes
 - Add missing skill fact type `StunBreak` as `Gw2Sharp.WebApi.V2.Models.SkillFactStunBreak` for pattern matching